🚚 rename timer.h -> time.h
