+++ dma/.vscode/c_cpp_properties.json
@@ -8,7 +8,7 @@
             "defines": [
                 "STM32F103xB"
             ],
-            "compilerPath": "/usr/bin/arm-none-eabi-gcc",
+            "compilerPath": "/Applications/ARM/bin/arm-none-eabi-gcc",
             "cStandard": "gnu17",
             "cppStandard": "gnu++17",
             "intelliSenseMode": "gcc-arm"
